Crate duk

A high-level wrapper around the Duktape Javascript/EcmaScript interpreter.

Currently, the focus is around supporting "extension"/"plug-in" use cases, so the primary supported functionality is:

Loading code.

Calling functions and getting their result.

Other use-cases (like exposing Rust functions to JS) are not yet implemented.

Structs

Context
A context corresponding to a thread of script execution.

ContextBuilder

Error

JsError
An error that originates from executing Javascript/Ecmascript.

Reference
A reference to a value that lives within a Context.

Enums

ErrorKind

JsErrorKind
Kinds of Javascript/Ecmascript errors

Value
A Javascript/Ecmascript value that exists in the Rust world.

Traits

Argument
Something that can be used as an argument when calling into Javascript code.

ChainErr

Type Definitions

ModuleLoader

ModuleResolver

Result