Enum duk::Value

pub enum Value {
    Undefined,
    Null,
    Boolean(bool),
    Number(f64),
    String(String),
    Array(Vec<Value>),
    Object(BTreeMap<String, Value>),
    Bytes(Vec<u8>),
    Foreign(&'static str),
}

A Javascript/Ecmascript value that exists in the Rust world.

Duktape supports values beyond these, but they don't have good Rust semantics, so they cannot be interacted with from the Rust world. They are therefore mapped to Value::Foreign when retrieved, and trying to further use those values is generally equivalent to using undefined.

Variants

Undefined

The undefined value.

Null

The null value.

Boolean

A boolean like true or false.

Number

Any number (both integral like 5 and fractional like 2.3).

String

Any string like 'abc'.

Array

Any array of values like ['a', 2, false].

Object

A JSON-like object like {a: 'a', b: 2, c: false}.

Bytes

A Duktape byte buffer like Duktape.Buffer('abc').

Foreign

A Duktape value that cannot be represented in Rust (yet).

Contains a &str describing the foreign type.

Methods

impl Value

fn to_reference<'a>(&self, context: &'a Context) -> Reference<'a>

Copies this value into a Context, and returns the reference to the value within the context.