Struct duk::Context

pub struct Context {
    // some fields omitted
}

A context corresponding to a thread of script execution.

Methods

impl Context

fn new() -> Context

Creates a new context.

fn builder() -> ContextBuilder

fn eval_string(&self, string: &str) -> Result<Reference>

Evaluates the specified script string within the current context.

Examples

Successful evaluation:

let mut ctx = duk::Context::new();
let value = ctx.eval_string("'ab' + 'cd' + Math.floor(2.3)").unwrap().to_value();
assert_eq!(duk::Value::String("abcd2".to_owned()), value);

However, if we try to call a function that doesn't exist:

let ctx = duk::Context::new();
let result = ctx.eval_string("var a = {}; a.foo()");
match result {
  Err(duk::Error(ref k, _)) => {
    match k {
      &duk::ErrorKind::Js(duk::JsError { kind, ref message, .. }) => {
        assert_eq!(duk::JsErrorKind::Type, kind);
        assert_eq!("undefined not callable", message);
      },
      _ => unreachable!(),
    }
  },
  _ => unreachable!(),
}

fn eval_string_with_filename(&self, filename: &str, string: &str) -> Result<Reference>

Like eval_string, but sets the file name for all of the evaluated functions to the specified string.

fn eval_file(&self, path: &Path) -> Result<Reference>

Loads and evaluates the specified file within the current context.

fn global_object(&self) -> Reference

Retrieves a reference to the global object.

fn call_global(&self, name: &str, args: &[&Argument]) -> Result<Reference>

Calls the specified global script function with the supplied arguments.

Behaves like global_object().call_method(name, args).