Struct duk::Reference

pub struct Reference<'a> {
    // some fields omitted
}

A reference to a value that lives within a Context.

Methods

impl<'a> Reference<'a>

fn to_value(&self) -> Value

Converts this reference to a Value which can be used for further processing by Rust code.

fn get(&self, name: &str) -> Result<Reference<'a>>

Gets the property with the specified key, provided that this reference points to something that is object coercible.

fn call(&self, args: &[&Argument]) -> Result<Reference<'a>>

Calls the function that this reference points to without a this binding, using the specified arguments.

When the function executes, the this binding is set to undefined or the global object, depending on if the function is strict or not. Calling this function is equivalent to doing myfunc.call(undefined, args) in Javascript.

fn call_with_this(&self, this: &Argument, args: &[&Argument]) -> Result<Reference<'a>>

Calls the function that this reference points to with an explicit this binding.

fn call_method(&self, name: &str, args: &[&Argument]) -> Result<Reference<'a>>

Calls a method on the object that this reference points to.

The this binding will be set to the object during the execution of the function. Calling this function is equivalent to doing myobj[name](args...) in Javascript.

fn new(&self, args: &[&Argument]) -> Result<Reference<'a>>

Calls the function that this reference points to as a constructor, with the specified arguments.