Trait duk::Argument

pub trait Argument {
    unsafe fn push_to_context(&self, context: &Context);
}

Something that can be used as an argument when calling into Javascript code.

Required Methods

unsafe fn push_to_context(&self, context: &Context)

Pushes this argument to the stack of the specified context. This requires interaction with the internals of the context, and is therefore an unsafe operation.